+++ src/ptk/ptk-file-task.cxx
@@ -1458,7 +1458,7 @@ ptk::file_task::update() noexcept
         i32 ipercent = 0;
         if (this->task->total_size)
         {
-            ipercent = (this->task->progress * 100) / this->task->total_size;
+            ipercent = static_cast<std::int32_t>((this->task->progress * 100) / this->task->total_size);
         }
         else
         {
@@ -2524,7 +2524,9 @@ ptk::file_task::query_overwrite() noexcept
     g_signal_connect(G_OBJECT(query_input), "key-press-event", G_CALLBACK(on_query_input_keypress), this);
     // clang-format on
     GtkWidget* input_buf = GTK_WIDGET(gtk_text_view_get_buffer(GTK_TEXT_VIEW(query_input)));
-    gtk_text_buffer_get_iter_at_offset(GTK_TEXT_BUFFER(input_buf), &iter, pos);
+    gtk_text_buffer_get_iter_at_offset(GTK_TEXT_BUFFER(input_buf),
+                                       &iter,
+                                       static_cast<std::int32_t>(pos));
     gtk_text_buffer_place_cursor(GTK_TEXT_BUFFER(input_buf), &iter);
     // clang-format off
     g_signal_connect(G_OBJECT(input_buf), "changed", G_CALLBACK(on_multi_input_changed), query_input);
